Serialize snippet node locations.
Snippet node locations are now saved in a separate file called 'GRAPH' located in the .snippet folder. When a snippet is opened, all lua files are retrieved and a snippet
is created for each lua/ut.lua pair of files. Then, the 'GRAPH' file is loaded to update each of the snippet positions.