Runtime: Snippet temperature animation.
This commit adds support for visualizing how recent a snippet was executed. An animation is played to interpolate the border color from hot to cold to transparent.
The 'BoxInstance' field has been deprecated in favor of overriding the 'Box' instance with a duplicate for animations to be played on a per-instance basis instead of per-resource.
Runtime singleton that is the central location for managing running all connected snippets. It also emits a signal for state changes and when a snippet is executed.