Snippet refactor to move data into its own object.
This commit moves data that used to be mutated in the Snippet object to the SnippetData object.
The Workspace now manages this list of SnippetData objects which the visual node holds a reference to.
Various components of the UI system had to change to reflect how the new SnippetData object is managed.