Run snippets through command-line.
Snippets can now be run through the command-line.
The '--snippet-run' option specifies the workspace to load.
The '--snippet-name' option specifies which snippet to run. If none is specified, then the 'main' snippet is run.
The '--snippet-log' option saves the output to a separate log file to prevent conflicts with the default Godot log system.