LuaCompilerResult class used for returning information about compilation.
LuaCompilerResult will contain the same information as LuaResult, but adds symbol
names to the object for use in the application.

The symbols are retrieved by iterating over all fields of a table and getting names of
values that are booleans, numbers, strings, functions, and tables. The '_G' table is
ignored along with tables named as 'loaded'. This is to prevent any circular
dependencies and causing the iteration to never halt.