Store line break for clients to be queried when re-opening snippet window.
When a debugging session breaks, the client is notified so that the UI
properly reflects which snippet and the line number the execution is halted
on. If the snippet window is closed and re-opened, the information is not
restored. The debugger now stores this information so that it can be
retrieved.

The Runtime object needed to be updated to update the state of the remote
task. Additional functions were added to properly handle this as well as
emit the proper signals when necessary.

Close #14.