Autocomplete now parses member operators.
The autocomplete function will now parse the member operator if one exists and filter out the possible keywords based off of the parent scope of the operator.

The function GetWordAtCursor has been updated to use a RegEx object to properly
find the word under the cursor. This will include all characters except for spaces.

The SetWordAtCursor has been replaced with the function ReplaceWordAtCursor which will use the TextEdit API to search for the word to replace only on the cursor
line. These changes allow for GetWordAtCursorStart to be deprecated.

Close #4.