Dynamically add/remove output pins with function name labels.
The Pin scene was reworked to remove scaling on the root node and moved to the texture rendering.
This scene also adds a Label2D node for displaying text next to the pin. The scaling that used to
be on the root node of the Pin scene is now applied on the PinConnection scene.

The Snippet graph node handles dynamically adding and removing functions that
have been defined in the code.