Added MaxWidth editable property to manage visual length of text.
The Label2D node can now clamp what the visual text will render. If
the text exceeds the max width, then it is truncated and a '...' is
added to the end to indicate more exists.

The Pin scene has been updated to apply a max width for all pins.
The offset of the Name is updated to reflect the new visual text.